Unify CoreML and Metal into single Metal backend
Consolidate the previously separate CoreML and MPSGraph backends into a
unified Metal backend. The hybrid architecture (CPU+GPU+ANE) is preserved
but exposed as a single -DUSE_BACKEND=METAL build option.

Key changes:
- Remove standalone CoreML backend files (coremlbackend.{cpp,h,swift})
- Merge CoreML logic into metalbackend.{cpp,h,swift}
- Rename mpsgraphlayers.swift to metallayers.swift
- Rename CoreMLProcess namespace to MetalProcess
- Update CMakeLists.txt: remove COREML backend, keep METAL only
- Update CI workflow to use METAL backend
- Prefer MPSGraph over CoreML for batch size 1 (better latency)
- Add autoreleasepool to CoreML dispatch path
- Remove unused alpha constant from ThroughputTracker
- Fix minor comment and log prefix inconsistencies
